feat: add download api key page (#367)
* feat: add download api key page

* feat: add routing changes

* feat: add a login redirect

* feat: remove download access key components

* feat: download the access key after the documents upload

* feat: get api user id from the user

---------
