+++ src/app/view/register/register.component.html
@@ -367,9 +367,6 @@
         </form>
       </mat-step>
     </mat-stepper>
-    <div *ngIf="showPopup" class="popup-container">
-      <div class="popup">please wait, Access key Downloading...</div>
-    </div>
   </mat-card-content>
 
   <div class="d-flex justify-content-between mb-2 px-5">
